Update sample with anchor and bullet-letter plugins
